requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
httpx==0.28.0
supabase==2.10.0
anthropic==0.40.0
apify-client==1.9.0
resend==2.5.0
weasyprint==63.1
jinja2==3.1.4
streamlit==1.41.0
python-dotenv==1.0.1
python-multipart==0.0.17
